multi-agent-system/pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langgraph>=0.6.6",
    "openai>=1.101.0",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
